+++ src/main/java/ru/javawebinar/topjava/model/MealTo.java
@@ -11,11 +11,14 @@ public class MealTo {
 
     private final boolean excess;
 
+    private final String color;
+
     public MealTo(LocalDateTime dateTime, String description, int calories, boolean excess) {
         this.dateTime = dateTime;
         this.description = description;
         this.calories = calories;
         this.excess = excess;
+        this.color = (excess) ? "red":"green";
     }
 
     @Override
@@ -25,6 +28,27 @@ public class MealTo {
                 ", description='" + description + '\'' +
                 ", calories=" + calories +
                 ", excess=" + excess +
+                ", color='" + color + '\'' +
                 '}';
+    }
+
+    public LocalDateTime getDateTime() {
+        return dateTime;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public int getCalories() {
+        return calories;
+    }
+
+    public boolean isExcess() {
+        return excess;
+    }
+
+    public String getColor() {
+        return color;
     }
 }